+++ src/main/java/me/anedhel/lotr/datagen/ModLootTableProvider.java
@@ -1,6 +1,6 @@
 package me.anedhel.lotr.datagen;
 
-import me.anedhel.lotr.blocks.ModBlocks;
+import me.anedhel.lotr.block.ModBlocks;
 import me.anedhel.lotr.item.ModItems;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricBlockLootTableProvider;
@@ -15,6 +15,8 @@ import net.minecraft.loot.entry.LootPoolEntry;
 import net.minecraft.loot.function.ApplyBonusLootFunction;
 import net.minecraft.loot.function.SetCountLootFunction;
 import net.minecraft.loot.provider.number.UniformLootNumberProvider;
+
+import java.util.List;
 
 /**
  * This class is used to add all LootTables the mod provides and edit existing ones.
@@ -37,6 +39,86 @@ public class ModLootTableProvider extends FabricBlockLootTableProvider {
         addDrop(ModBlocks.TIN_BLOCK);
         addDrop(ModBlocks.BRONZE_BLOCK);
 
+        addDrop(ModBlocks.ANDESITE_BUTTON);
+        addDrop(ModBlocks.ANDESITE_PRESSURE_PLATE);
+
+        addDrop(ModBlocks.POLISHED_ANDESITE_WALL);
+        addDrop(ModBlocks.POLISHED_ANDESITE_BUTTON);
+        addDrop(ModBlocks.POLISHED_ANDESITE_PRESSURE_PLATE);
+
+        addDrop(ModBlocks.COBBLED_ANDESITE);
+        addDrop(ModBlocks.COBBLED_ANDESITE_STAIRS);
+        addDrop(ModBlocks.COBBLED_ANDESITE_WALL);
+        addDrop(ModBlocks.COBBLED_ANDESITE_BUTTON);
+        addDrop(ModBlocks.COBBLED_ANDESITE_PRESSURE_PLATE);
+        addDrop(ModBlocks.COBBLED_ANDESITE_SLAB, slabDrops(ModBlocks.COBBLED_ANDESITE_SLAB));
+
+        addDrop(ModBlocks.ANDESITE_BRICK);
+        addDrop(ModBlocks.ANDESITE_BRICK_STAIRS);
+        addDrop(ModBlocks.ANDESITE_BRICK_WALL);
+        addDrop(ModBlocks.ANDESITE_BRICK_SLAB, slabDrops(ModBlocks.ANDESITE_BRICK_SLAB));
+
+        addDrop(ModBlocks.DIORITE_BUTTON);
+        addDrop(ModBlocks.DIORITE_PRESSURE_PLATE);
+
+        addDrop(ModBlocks.POLISHED_DIORITE_WALL);
+        addDrop(ModBlocks.POLISHED_DIORITE_BUTTON);
+        addDrop(ModBlocks.POLISHED_DIORITE_PRESSURE_PLATE);
+
+        addDrop(ModBlocks.COBBLED_DIORITE);
+        addDrop(ModBlocks.COBBLED_DIORITE_STAIRS);
+        addDrop(ModBlocks.COBBLED_DIORITE_WALL);
+        addDrop(ModBlocks.COBBLED_DIORITE_BUTTON);
+        addDrop(ModBlocks.COBBLED_DIORITE_PRESSURE_PLATE);
+        addDrop(ModBlocks.COBBLED_DIORITE_SLAB, slabDrops(ModBlocks.COBBLED_DIORITE_SLAB));
+
+        addDrop(ModBlocks.DIORITE_BRICK);
+        addDrop(ModBlocks.DIORITE_BRICK_STAIRS);
+        addDrop(ModBlocks.DIORITE_BRICK_WALL);
+        addDrop(ModBlocks.DIORITE_BRICK_SLAB, slabDrops(ModBlocks.DIORITE_BRICK_SLAB));
+
+        addDrop(ModBlocks.GRANITE_BUTTON);
+        addDrop(ModBlocks.GRANITE_PRESSURE_PLATE);
+
+        addDrop(ModBlocks.POLISHED_GRANITE_WALL);
+        addDrop(ModBlocks.POLISHED_GRANITE_BUTTON);
+        addDrop(ModBlocks.POLISHED_GRANITE_PRESSURE_PLATE);
+
+        addDrop(ModBlocks.COBBLED_GRANITE);
+        addDrop(ModBlocks.COBBLED_GRANITE_STAIRS);
+        addDrop(ModBlocks.COBBLED_GRANITE_WALL);
+        addDrop(ModBlocks.COBBLED_GRANITE_BUTTON);
+        addDrop(ModBlocks.COBBLED_GRANITE_PRESSURE_PLATE);
+        addDrop(ModBlocks.COBBLED_GRANITE_SLAB, slabDrops(ModBlocks.COBBLED_GRANITE_SLAB));
+
+        addDrop(ModBlocks.GRANITE_BRICK);
+        addDrop(ModBlocks.GRANITE_BRICK_STAIRS);
+        addDrop(ModBlocks.GRANITE_BRICK_WALL);
+        addDrop(ModBlocks.GRANITE_BRICK_SLAB, slabDrops(ModBlocks.GRANITE_BRICK_SLAB));
+
+        addDrop(ModBlocks.SMOOTH_BASALT_STAIRS);
+        addDrop(ModBlocks.SMOOTH_BASALT_SLAB, slabDrops(ModBlocks.SMOOTH_BASALT_SLAB));
+        addDrop(ModBlocks.SMOOTH_BASALT_WALL);
+        addDrop(ModBlocks.SMOOTH_BASALT_BUTTON);
+        addDrop(ModBlocks.SMOOTH_BASALT_PRESSURE_PLATE);
+
+        addDrop(ModBlocks.BASALT_BRICK);
+        addDrop(ModBlocks.BASALT_BRICK_STAIRS);
+        addDrop(ModBlocks.BASALT_BRICK_SLAB, slabDrops(ModBlocks.BASALT_BRICK_SLAB));
+        addDrop(ModBlocks.BASALT_BRICK_WALL);
+
+        addDrop(ModBlocks.BLUESLATE);
+        addDrop(ModBlocks.BLUESLATE_STAIRS);
+        addDrop(ModBlocks.BLUESLATE_SLAB);
+        addDrop(ModBlocks.BLUESLATE_WALL);
+        addDrop(ModBlocks.BLUESLATE_BUTTON);
+        addDrop(ModBlocks.BLUESLATE_PRESSURE_PLATE);
+
+        addDrop(ModBlocks.BLUESLATE_BRICK);
+        addDrop(ModBlocks.BLUESLATE_BRICK_STAIRS);
+        addDrop(ModBlocks.BLUESLATE_BRICK_SLAB);
+        addDrop(ModBlocks.BLUESLATE_BRICK_WALL);
+
         addDrop(ModBlocks.CHALK);
         addDrop(ModBlocks.CHALK_STAIRS);
         addDrop(ModBlocks.CHALK_WALL);
@@ -47,6 +129,7 @@ public class ModLootTableProvider extends FabricBlockLootTableProvider {
         addDrop(ModBlocks.CHALK_BRICK);
         addDrop(ModBlocks.CHALK_BRICK_STAIRS);
         addDrop(ModBlocks.CHALK_BRICK_WALL);
+        addDrop(ModBlocks.CHALK_BRICK_SLAB, slabDrops(ModBlocks.CHALK_BRICK_SLAB));
 
         addDrop(ModBlocks.COBBLED_CHALK);
         addDrop(ModBlocks.COBBLED_CHALK_STAIRS);