+++ src/main/java/me/anedhel/lotr/datagen/ModModelProvider.java
@@ -1,13 +1,17 @@
 package me.anedhel.lotr.datagen;
 
-import me.anedhel.lotr.blocks.ModBlocks;
+import me.anedhel.lotr.block.ModBlocks;
+import me.anedhel.lotr.block.ModStoneType;
+import me.anedhel.lotr.item.ModGearType;
 import me.anedhel.lotr.item.ModItems;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricModelProvider;
+import net.minecraft.block.Block;
 import net.minecraft.data.client.BlockStateModelGenerator;
 import net.minecraft.data.client.ItemModelGenerator;
 import net.minecraft.data.client.Models;
 import net.minecraft.item.ArmorItem;
+import net.minecraft.item.Item;
 
 /**
  * This class is used to generate all models the mod provides.
@@ -29,30 +33,30 @@ public class ModModelProvider extends FabricModelProvider {
         blockStateModelGenerator.registerSimpleCubeAll(ModBlocks.TIN_BLOCK);
         blockStateModelGenerator.registerSimpleCubeAll(ModBlocks.BRONZE_BLOCK);
 
-        BlockStateModelGenerator.BlockTexturePool chalkPool = blockStateModelGenerator
-                .registerCubeAllModelTexturePool(ModBlocks.CHALK);
+        generateModStoneTypeModels(ModStoneType.ANDESITE, blockStateModelGenerator);
+        generateModStoneTypeModels(ModStoneType.POLISHED_ANDESITE, blockStateModelGenerator);
+        generateModStoneTypeModels(ModStoneType.COBBLED_ANDESITE, blockStateModelGenerator);
+        generateModStoneTypeModels(ModStoneType.ANDESITE_BRICK, blockStateModelGenerator);
 
-        chalkPool.stairs(ModBlocks.CHALK_STAIRS);
-        chalkPool.slab(ModBlocks.CHALK_SLAB);
-        chalkPool.wall(ModBlocks.CHALK_WALL);
-        chalkPool.button(ModBlocks.CHALK_BUTTON);
-        chalkPool.pressurePlate(ModBlocks.CHALK_PRESSURE_PLATE);
+        generateModStoneTypeModels(ModStoneType.DIORITE, blockStateModelGenerator);
+        generateModStoneTypeModels(ModStoneType.POLISHED_DIORITE, blockStateModelGenerator);
+        generateModStoneTypeModels(ModStoneType.COBBLED_DIORITE, blockStateModelGenerator);
+        generateModStoneTypeModels(ModStoneType.DIORITE_BRICK, blockStateModelGenerator);
 
-        BlockStateModelGenerator.BlockTexturePool chalkBrickPool = blockStateModelGenerator
-                .registerCubeAllModelTexturePool(ModBlocks.CHALK_BRICK);
+        generateModStoneTypeModels(ModStoneType.GRANITE, blockStateModelGenerator);
+        generateModStoneTypeModels(ModStoneType.POLISHED_GRANITE, blockStateModelGenerator);
+        generateModStoneTypeModels(ModStoneType.COBBLED_GRANITE, blockStateModelGenerator);
+        generateModStoneTypeModels(ModStoneType.GRANITE_BRICK, blockStateModelGenerator);
 
-        chalkBrickPool.stairs(ModBlocks.CHALK_BRICK_STAIRS);
-        chalkBrickPool.slab(ModBlocks.CHALK_BRICK_SLAB);
-        chalkBrickPool.wall(ModBlocks.CHALK_BRICK_WALL);
+        generateModStoneTypeModels(ModStoneType.SMOOTH_BASALT, blockStateModelGenerator);
+        generateModStoneTypeModels(ModStoneType.BASALT_BRICKS, blockStateModelGenerator);
 
-        BlockStateModelGenerator.BlockTexturePool cobblesChalkPool = blockStateModelGenerator
-                .registerCubeAllModelTexturePool(ModBlocks.COBBLED_CHALK);
+        generateModStoneTypeModels(ModStoneType.BLUESLATE, blockStateModelGenerator);
+        generateModStoneTypeModels(ModStoneType.BLUESLATE_BRICK, blockStateModelGenerator);
 
-        cobblesChalkPool.stairs(ModBlocks.COBBLED_CHALK_STAIRS);
-        cobblesChalkPool.slab(ModBlocks.COBBLED_CHALK_SLAB);
-        cobblesChalkPool.wall(ModBlocks.COBBLED_CHALK_WALL);
-        cobblesChalkPool.button(ModBlocks.COBBLED_CHALK_BUTTON);
-        cobblesChalkPool.pressurePlate(ModBlocks.COBBLED_CHALK_PRESSURE_PLATE);
+        generateModStoneTypeModels(ModStoneType.CHALK, blockStateModelGenerator);
+        generateModStoneTypeModels(ModStoneType.COBBLED_CHALK, blockStateModelGenerator);
+        generateModStoneTypeModels(ModStoneType.CHALK_BRICK, blockStateModelGenerator);
     }
 
     /**
@@ -65,15 +69,86 @@ public class ModModelProvider extends FabricModelProvider {
         itemModelGenerator.register(ModItems.TIN_INGOT, Models.GENERATED);
         itemModelGenerator.register(ModItems.BRONZE_INGOT, Models.GENERATED);
 
-        itemModelGenerator.register(ModItems.BRONZE_SWORD, Models.HANDHELD);
-        itemModelGenerator.register(ModItems.BRONZE_PICKAXE, Models.HANDHELD);
-        itemModelGenerator.register(ModItems.BRONZE_AXE, Models.HANDHELD);
-        itemModelGenerator.register(ModItems.BRONZE_SHOVEL, Models.HANDHELD);
-        itemModelGenerator.register(ModItems.BRONZE_HOE, Models.HANDHELD);
+        generateModGearTypeModels(ModGearType.BRONZE, itemModelGenerator);
+    }
 
-        itemModelGenerator.registerArmor((ArmorItem) ModItems.BRONZE_HELMET);
-        itemModelGenerator.registerArmor((ArmorItem) ModItems.BRONZE_CHESTPLATE);
-        itemModelGenerator.registerArmor((ArmorItem) ModItems.BRONZE_LEGGINGS);
-        itemModelGenerator.registerArmor((ArmorItem) ModItems.BRONZE_BOOTS);
+    /**
+     * This Method is used to generate all the Models for the given GearType
+     * @param gearType The ModGearType the models should be generated for.
+     * @param itemModelGenerator an ItemModelGenerator provided by minecraft
+     */
+    private void generateModGearTypeModels(ModGearType gearType, ItemModelGenerator itemModelGenerator) {
+        Item helmet = gearType.getHelmet();
+        Item chestplate = gearType.getChestplate();
+        Item leggings = gearType.getLeggings();
+        Item boots = gearType.getBoots();
+        Item sword = gearType.getSword();
+        Item axe = gearType.getAxe();
+        Item pickaxe = gearType.getPickaxe();
+        Item shovel = gearType.getShovel();
+        Item hoe = gearType.getHoe();
+
+        if (helmet != null) {
+            itemModelGenerator.registerArmor((ArmorItem) ModItems.BRONZE_HELMET);
+        }
+        if (chestplate != null) {
+            itemModelGenerator.registerArmor((ArmorItem) ModItems.BRONZE_CHESTPLATE);
+        }
+        if (leggings != null) {
+            itemModelGenerator.registerArmor((ArmorItem) ModItems.BRONZE_LEGGINGS);
+        }
+        if (boots != null) {
+            itemModelGenerator.registerArmor((ArmorItem) ModItems.BRONZE_BOOTS);
+        }
+        if (sword != null) {
+            itemModelGenerator.register(ModItems.BRONZE_SWORD, Models.HANDHELD);
+        }
+        if (axe != null) {
+            itemModelGenerator.register(ModItems.BRONZE_AXE, Models.HANDHELD);
+        }
+        if (pickaxe != null) {
+            itemModelGenerator.register(ModItems.BRONZE_PICKAXE, Models.HANDHELD);
+        }
+        if (shovel != null) {
+            itemModelGenerator.register(ModItems.BRONZE_SHOVEL, Models.HANDHELD);
+        }
+        if (hoe != null) {
+            itemModelGenerator.register(ModItems.BRONZE_HOE, Models.HANDHELD);
+        }
+    }
+
+    /**
+     * This Method is used to generate all the models for the given stone type.
+     * It only works for blocks completely added by the mod, adding Models for additions made based on vanilla blocks does not work with this method.
+     * @param stoneType The ModStoneType, for which the models should be created
+     * @param blockStateModelGenerator a BlockStateModelGenerator provided by minecraft
+     */
+    private void generateModStoneTypeModels(ModStoneType stoneType, BlockStateModelGenerator blockStateModelGenerator) {
+        Block baseBlock = stoneType.getBaseBlock();
+        Block stairBlock = stoneType.getStairBlock();
+        Block slabBlock = stoneType.getSlabBlock();
+        Block wallBlock = stoneType.getWallBlock();
+        Block buttonBlock = stoneType.getButtonBlock();
+        Block pressurePlateBlock = stoneType.getPressurePlateBlock();
+
+        if (baseBlock != null) {
+            BlockStateModelGenerator.BlockTexturePool stoneTypePool = blockStateModelGenerator
+                    .registerCubeAllModelTexturePool(baseBlock);
+            if (stairBlock != null) {
+                stoneTypePool.stairs(stairBlock);
+            }
+            if (slabBlock != null) {
+                stoneTypePool.slab(slabBlock);
+            }
+            if (wallBlock != null) {
+                stoneTypePool.wall(wallBlock);
+            }
+            if (buttonBlock != null) {
+                stoneTypePool.button(buttonBlock);
+            }
+            if (pressurePlateBlock != null) {
+                stoneTypePool.pressurePlate(pressurePlateBlock);
+            }
+        }
     }
 }