+++ src/main/java/me/anedhel/lotr/item/ModItems.java
@@ -12,13 +12,6 @@ import net.minecraft.util.Identifier;
  */
 public class ModItems {
 
-
-    /**
-     * This Method is used to register Items with minecraft.
-     * @param name the name of the Item
-     * @param item the Item
-     * @return the Item after it has been registered
-     */
     public static final Item RAW_TIN = registerItem("raw_tin", new Item(new FabricItemSettings()));
     public static final Item TIN_INGOT = registerItem("tin_ingot", new Item(new FabricItemSettings()));
     public static final Item BRONZE_INGOT = registerItem("bronze_ingot", new Item(new FabricItemSettings()));
@@ -35,15 +28,21 @@ public class ModItems {
             new HoeItem(ModToolMaterial.BRONZE, -2, -1.0f, new FabricItemSettings()));
 
     public static final Item BRONZE_HELMET = registerItem("bronze_helmet",
-            new ArmorItem(ModArmorMaterials.BRONZE, ArmorItem.Type.HELMET, new FabricItemSettings()));
+            new ArmorItem(ModArmorMaterial.BRONZE, ArmorItem.Type.HELMET, new FabricItemSettings()));
     public static final Item BRONZE_CHESTPLATE = registerItem("bronze_chestplate",
-            new ArmorItem(ModArmorMaterials.BRONZE, ArmorItem.Type.CHESTPLATE, new FabricItemSettings()));
+            new ArmorItem(ModArmorMaterial.BRONZE, ArmorItem.Type.CHESTPLATE, new FabricItemSettings()));
     public static final Item BRONZE_LEGGINGS = registerItem("bronze_leggings",
-            new ArmorItem(ModArmorMaterials.BRONZE, ArmorItem.Type.LEGGINGS, new FabricItemSettings()));
+            new ArmorItem(ModArmorMaterial.BRONZE, ArmorItem.Type.LEGGINGS, new FabricItemSettings()));
     public static final Item BRONZE_BOOTS = registerItem("bronze_boots",
-            new ArmorItem(ModArmorMaterials.BRONZE, ArmorItem.Type.BOOTS, new FabricItemSettings()));
+            new ArmorItem(ModArmorMaterial.BRONZE, ArmorItem.Type.BOOTS, new FabricItemSettings()));
 
 
+    /**
+     * This Method is used to register Items with minecraft.
+     * @param name the name of the Item
+     * @param item the Item
+     * @return the Item after it has been registered
+     */
     private static Item registerItem(String name, Item item) {
         return Registry.register(Registries.ITEM,
                 new Identifier(LordOfTheRingsMiddleEarthMod.MOD_ID, name), item);