+++ CHANGELOG.md
@@ -8,3 +8,6 @@ ___
 + Added Tin Ore, Raw Tin and Tin Ingots with their CompactingBlocks
 + Added Bronze Ingot and Bronze Block
 + Added Bronze Equipment (Armor, Sword, Axe, Pickaxe, Shovel, Hoe) 
++ Added Andesite Button and Pressure Plate
++ Added Cobbled Andesite with it´s building blocks
++ Added Andesite Bricks with it´s building blocks